+++ src/main/java/graphql/execution/AsyncExecutionStrategy.java
@@ -1,30 +1,13 @@
 package graphql.execution;
 
-import graphql.ExecutionResult;
-import graphql.execution.defer.DeferSupport;
-import graphql.execution.defer.DeferredCall;
-import graphql.execution.defer.DeferredErrorSupport;
-import graphql.execution.instrumentation.DeferredFieldInstrumentationContext;
-import graphql.execution.instrumentation.ExecutionStrategyInstrumentationContext;
-import graphql.execution.instrumentation.Instrumentation;
-import graphql.execution.instrumentation.parameters.InstrumentationDeferredFieldParameters;
-import graphql.execution.instrumentation.parameters.InstrumentationExecutionStrategyParameters;
-import graphql.language.Field;
-import graphql.schema.GraphQLFieldDefinition;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.function.BiConsumer;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
+import graphql.PublicApi;
 
 /**
  * The standard graphql execution strategy that runs fields asynchronously non-blocking.
  */
+@PublicApi
 public class AsyncExecutionStrategy extends AbstractAsyncExecutionStrategy {
+
 
     /**
      * The standard graphql execution strategy that runs fields asynchronously
@@ -42,98 +25,4 @@ public class AsyncExecutionStrategy extends AbstractAsyncExecutionStrategy {
         super(exceptionHandler);
     }
 
-    @Override
-    @SuppressWarnings("FutureReturnValueIgnored")
-    public CompletableFuture<ExecutionResult> execute(ExecutionContext executionContext, ExecutionStrategyParameters parameters) throws NonNullableFieldWasNullException {
-
-        Instrumentation instrumentation = executionContext.getInstrumentation();
-        InstrumentationExecutionStrategyParameters instrumentationParameters = new InstrumentationExecutionStrategyParameters(executionContext, parameters);
-
-        ExecutionStrategyInstrumentationContext executionStrategyCtx = instrumentation.beginExecutionStrategy(instrumentationParameters);
-
-        Map<String, List<Field>> fields = parameters.getFields();
-        List<String> fieldNames = new ArrayList<>(fields.keySet());
-        List<CompletableFuture<FieldValueInfo>> futures = new ArrayList<>();
-        List<String> resolvedFields = new ArrayList<>();
-        for (String fieldName : fieldNames) {
-            List<Field> currentField = fields.get(fieldName);
-
-            ExecutionPath fieldPath = parameters.getPath().segment(fieldName);
-            ExecutionStrategyParameters newParameters = parameters
-                    .transform(builder -> builder.field(currentField).path(fieldPath).parent(parameters));
-
-            if (isDeferred(executionContext, newParameters, currentField)) {
-                executionStrategyCtx.onDeferredField(currentField);
-                continue;
-            }
-            resolvedFields.add(fieldName);
-            CompletableFuture<FieldValueInfo> future = resolveFieldWithInfo(executionContext, newParameters);
-            futures.add(future);
-        }
-        CompletableFuture<ExecutionResult> overallResult = new CompletableFuture<>();
-        executionStrategyCtx.onDispatched(overallResult);
-
-        Async.each(futures).whenComplete((completeValueInfos, throwable) -> {
-            BiConsumer<List<ExecutionResult>, Throwable> handleResultsConsumer = handleResults(executionContext, resolvedFields, overallResult);
-            if (throwable != null) {
-                handleResultsConsumer.accept(null, throwable.getCause());
-                return;
-            }
-            List<CompletableFuture<ExecutionResult>> executionResultFuture = completeValueInfos.stream().map(FieldValueInfo::getFieldValue).collect(Collectors.toList());
-            executionStrategyCtx.onFieldValuesInfo(completeValueInfos);
-            Async.each(executionResultFuture).whenComplete(handleResultsConsumer);
-        });
-
-        overallResult.whenComplete(executionStrategyCtx::onCompleted);
-        return overallResult;
-    }
-
-    private boolean isDeferred(ExecutionContext executionContext, ExecutionStrategyParameters parameters, List<Field> currentField) {
-        DeferSupport deferSupport = executionContext.getDeferSupport();
-        if (deferSupport.checkForDeferDirective(currentField)) {
-            DeferredErrorSupport errorSupport = new DeferredErrorSupport();
-
-            // with a deferred field we are really resetting where we execute from, that is from this current field onwards
-            Map<String, List<Field>> fields = new HashMap<>();
-            fields.put(currentField.get(0).getName(), currentField);
-
-            ExecutionStrategyParameters callParameters = parameters.transform(builder ->
-                    builder.deferredErrorSupport(errorSupport)
-                            .field(currentField)
-                            .fields(fields)
-                            .parent(null) // this is a break in the parent -> child chain - its a new start effectively
-                            .listSize(0)
-                            .currentListIndex(0)
-            );
-
-            DeferredCall call = new DeferredCall(deferredExecutionResult(executionContext, callParameters), errorSupport);
-            deferSupport.enqueue(call);
-            return true;
-        }
-        return false;
-    }
-
-    @SuppressWarnings("FutureReturnValueIgnored")
-    private Supplier<CompletableFuture<ExecutionResult>> deferredExecutionResult(ExecutionContext executionContext, ExecutionStrategyParameters parameters) {
-        return () -> {
-            GraphQLFieldDefinition fieldDef = getFieldDef(executionContext, parameters, parameters.getField().get(0));
-
-            Instrumentation instrumentation = executionContext.getInstrumentation();
-            DeferredFieldInstrumentationContext fieldCtx = instrumentation.beginDeferredField(
-                    new InstrumentationDeferredFieldParameters(executionContext, parameters, fieldDef, fieldTypeInfo(parameters, fieldDef))
-            );
-            CompletableFuture<ExecutionResult> result = new CompletableFuture<>();
-            fieldCtx.onDispatched(result);
-            CompletableFuture<FieldValueInfo> fieldValueInfoFuture = resolveFieldWithInfo(executionContext, parameters);
-
-            fieldValueInfoFuture.whenComplete((fieldValueInfo, throwable) -> {
-                fieldCtx.onFieldValueInfo(fieldValueInfo);
-
-                CompletableFuture<ExecutionResult> execResultFuture = fieldValueInfo.getFieldValue();
-                execResultFuture = execResultFuture.whenComplete(fieldCtx::onCompleted);
-                Async.copyResults(execResultFuture, result);
-            });
-            return result;
-        };
-    }
 }
