+++ src/main/java/graphql/execution/AsyncSerialExecutionStrategy.java
@@ -1,54 +1,20 @@
 package graphql.execution;
 
-import graphql.ExecutionResult;
-import graphql.execution.instrumentation.Instrumentation;
-import graphql.execution.instrumentation.InstrumentationContext;
-import graphql.execution.instrumentation.parameters.InstrumentationExecutionStrategyParameters;
-import graphql.language.Field;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
+import graphql.PublicApi;
 
 /**
  * Async non-blocking execution, but serial: only one field at the the time will be resolved.
  * See {@link AsyncExecutionStrategy} for a non serial (parallel) execution of every field.
  */
+@PublicApi
 public class AsyncSerialExecutionStrategy extends AbstractAsyncExecutionStrategy {
 
     public AsyncSerialExecutionStrategy() {
-        super(new SimpleDataFetcherExceptionHandler());
+        super(new SimpleDataFetcherExceptionHandler(), true);
     }
 
     public AsyncSerialExecutionStrategy(DataFetcherExceptionHandler exceptionHandler) {
-        super(exceptionHandler);
-    }
-
-    @Override
-    @SuppressWarnings({"TypeParameterUnusedInFormals","FutureReturnValueIgnored"})
-    public CompletableFuture<ExecutionResult> execute(ExecutionContext executionContext, ExecutionStrategyParameters parameters) throws NonNullableFieldWasNullException {
-
-        Instrumentation instrumentation = executionContext.getInstrumentation();
-        InstrumentationExecutionStrategyParameters instrumentationParameters = new InstrumentationExecutionStrategyParameters(executionContext, parameters);
-        InstrumentationContext<ExecutionResult> executionStrategyCtx = instrumentation.beginExecutionStrategy(instrumentationParameters);
-        Map<String, List<Field>> fields = parameters.getFields();
-        List<String> fieldNames = new ArrayList<>(fields.keySet());
-
-        CompletableFuture<List<ExecutionResult>> resultsFuture = Async.eachSequentially(fieldNames, (fieldName, index, prevResults) -> {
-            List<Field> currentField = fields.get(fieldName);
-            ExecutionPath fieldPath = parameters.getPath().segment(fieldName);
-            ExecutionStrategyParameters newParameters = parameters
-                    .transform(builder -> builder.field(currentField).path(fieldPath));
-            return resolveField(executionContext, newParameters);
-        });
-
-        CompletableFuture<ExecutionResult> overallResult = new CompletableFuture<>();
-        executionStrategyCtx.onDispatched(overallResult);
-
-        resultsFuture.whenComplete(handleResults(executionContext, fieldNames, overallResult));
-        overallResult.whenComplete(executionStrategyCtx::onCompleted);
-        return overallResult;
+        super(exceptionHandler, true);
     }
 
 }
